feat(shannon): to support notification mode for changes propagation
1: add new way of sync the changes to rapid.

Issues: #476
