+++ storage/innobase/srv/srv0srv.cc
@@ -81,7 +81,7 @@ this program; if not, write to the Free Software Foundation, Inc.,
 #include "sql/sql_class.h"
 #include "sql_thd_internal_api.h"
 #include "srv0mon.h"
-#include "storage/rapid_engine/populate/populate.h"
+#include "storage/rapid_engine/populate/log_commons.h"
 
 #include "my_dbug.h"
 #include "my_psi_config.h"
