+++ storage/innobase/srv/srv0start.cc
@@ -126,7 +126,7 @@ this program; if not, write to the Free Software Foundation, Inc.,
 #include "ut0crc32.h"
 #include "ut0new.h"
 
-#include "storage/rapid_engine/populate/populate.h"
+#include "storage/rapid_engine/populate/log_commons.h"
 #include "storage/rapid_engine/imcs/purge/purge.h"
 
 /** fil_space_t::flags for hard-coded tablespaces */
